Update InvenTreeLink.py
Completely changed the way the plugin handles Template Parameters. These are now classes wrapped in an Enum, in the format of "Fusion360:" and another key, like "Area" or "Volume" which makes it "Fusion360:Volume" or "Fusion360: Area". Along with this the Parameter Templates are created when the plugin launches, so I removed the entry from 'conf.ini'

Renamed config keys for better understandability